+++ public/backend.js
@@ -15,7 +15,7 @@
 			USER_SETTINGS.language = 'en'
 
 		;[report_errors, bind_tooltip, bind_confirm, translate_calendar, onetime].forEach((f) => f.call())
-		;[page_dashboard, page_settings_main, page_user_pref, page_user_dashboard, page_bosmang]
+		;[page_dashboard, page_settings_main, page_user_pref, page_user_api, page_user_dashboard, page_bosmang]
 			.forEach((f) => document.body.id.match(new RegExp('^' + f.name.replace(/_/g, '-'))) && f.call())
 	})
 
@@ -184,6 +184,27 @@
 		$('[id="user.settings.fewer_numbers"]').on('change', function(e) {
 			$('#lock-settings').css('display', $(this).is(':checked') ? 'block' : 'none')
 		}).trigger('change')
+	}
+
+	let page_user_api = function() {
+		// Show API token
+		$('a[data-show]').on('click', (e) => {
+			e.preventDefault()
+			let elem = $(e.target)
+			elem.parent().text(elem.attr('data-show'))
+		})
+
+		// All sites
+		$('.allsites input').on('change', (e) => {
+			if (e.target.checked) {
+				$('.site').addClass('disabled')
+				$('.site input').attr('disabled', true).prop('checked', false)
+			} else {
+				$('.site').removeClass('disabled')
+				$('.site input').removeAttr('disabled')
+			}
+		})
+		$('.allsites input').trigger('change')
 	}
 
 	var page_user_dashboard = function() {